backend/requirements.txt
fastapi
uvicorn
google-generativeai
langchain
langchain-community
chromadb
sentence-transformers
reportlab
python-docx
sarvam-client
